Use `notificationUuid` to load notification in editor
Instead of directly passing the notification to the editor screen, we will pass the notification uuid to load the notification. Once loaded we can update the model to have the notification. In later commits we will use this notification to render the parts of UI instead of using the arguments provided notification. Once this is done, we can change the arguments option to notificationUuid.