Move text change listeners outside of view effects in EditorScreen
This is causing issues when screen is rotated, since set title and set content view effects are only fired when screen is created the text listeners are only attached then. It seems like when screen is rotated the text listeners are not being reattached. So moved the text listeners to onActivityCreated.