Use `AnimatedVectorDrawableCompat` in `SplashActivity`
`AnimatedVectorDrawable` uses different variants depending on the API version. For API 21 & 23 it used `AnimatedVectorDrawableCompat` where as for API 25 and above it used `AnimatedVectorDrawable`. Because I was casting the drawable gotten from the `ImageView` to `AnimatedVectorDrawable` it crashes on API 23. In order to fix that, I am constructing the AVD at runtime using `AnimatedVectorDrawableCompat`.

For more information:
https://medium.com/androiddevelopers/re-animation-7869722af206